add a nightly-only auto trait to prevent incremental nodes holding observers
This has been an implicit requirement all along. In practice, as a
codebase gets large, the likelihood of a someone putting huge state
objects that hold observers into incremental nodes approaches 1. It's
particularly annoying to debug because it only happens when the nodes
are dropped, which might not happen in regular usage of e.g. a
long-lived UI.

Incremental will throw errors if you try to *use* an observer. However,
it is also illegal to have them owned by e.g. a Map closure. Even if we
could prevent the IncrState from panicking trying to tear down parts of
its structure when this happens, it would still be valuable to prevent
anyone trying to get Observers' values inside an incremental node.

The new `incremental::incrsan::NotObserver` allows you to statically
check that you are not using Observers this way whatsoever. Switch to a
nightly compiler and enable the `nightly-incrsan` feature.